+++ CHANGES.rst
@@ -1,6 +1,7 @@
 0.9.2
 -----
 - Reverted to set() for versions of Python < 2.7 that lack set literal syntax
+- Removed keyword arguments from decode() to support versions of Python < 2.7
 - Fixed tests for servers like Lotus Domino that do not update search indices